requirements.txt
2captcha-python==1.1.3
alembic==1.8.1
async-generator==1.10
attrs==22.1.0
Automat==22.10.0
beautifulsoup4==4.11.1
blinker==1.5
cachelib==0.9.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
cssutils==2.6.0
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.2.1
docopt==0.6.2
emails==0.6
exceptiongroup==1.1.0
filelock==3.8.0
Flask==2.2.2
Flask-Caching==2.0.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gensim==3.8.1
greenlet==1.1.3
h11==0.14.0
helper==2.5.0
html2markdown==0.1.7
html2text==2020.1.16
hyperlink==21.0.0
idna==3.4
ImageHash==4.3.1
IMAPClient==2.3.1
imbox==0.9.8
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
lxml==4.9.1
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
nltk==3.8.1
numpy==1.24.2
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==2.0.1
parsel==1.7.0
pdfkit==1.0.0
Pillow==9.4.0
pipreqs==0.4.11
plotly==5.14.1
premailer==3.10.0
Protego==0.2.1
pyapollos==0.1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-frontmatter==1.0.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==3.13
qiniu==7.10.0
queuelib==1.6.2
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
revChatGPT==0.2.0
scipy==1.10.1
Scrapy==2.7.1
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
tenacity==8.2.2
tldextract==3.4.0
tls-client==0.1.8
tomd==0.1.3
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
tzdata==2023.3
undetected-chromedriver==3.2.1
unicode==2.9
urllib3==1.26.12
utils==1.0.1
w3lib==2.0.1
websockets==10.4
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yarg==0.1.9
zipp==3.8.1
zmail==0.2.0
zope.interface==5.5.1
